server/requirements.txt
pyautogui
yt-dlp
qrcode
pillow
win10toast
